+++ CHANGELOG.md
@@ -6,8 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 * Changed selector expected by `default strategy` in `move_solutions_to_answer_key` to optimize search (major)
-
 * Fix Rubocop GitHub Action's regular expression used to select files to lint (patch)
+* Added monkey patch for array to be able to add a prefix: `%w[multiple-choice true-false].prefix('section.')` (minor)
 
 ## [6.1.0] - 2021-06-21